% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_springer_books_available.R
\name{data_springer_books_available}
\alias{data_springer_books_available}
\title{Extract Data from Springer Database for Available Ebooks.}
\usage{
data_springer_books_available(type = c("pdf", "epub", "both"))
}
\arguments{
\item{type}{Character string.
Type of available ebooks.}
}
\description{
Extract Data from Springer Database for Available Ebooks.
}
\author{
Ivan Jacob Agaloos Pesigan
}
