% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_extract_isbn_crossref.R
\name{lib_extract_isbn_crossref}
\alias{lib_extract_isbn_crossref}
\title{Extract ISBN from the result of lib_metadata_crossref.}
\usage{
lib_extract_isbn_crossref(
  fromdoi = TRUE,
  doi = NULL,
  cf = NULL,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{fromdoi}{Logical.
If \code{TRUE}, uses the argument \code{doi}
and the function \code{lib_metadata_crossref}
to generate ISBN.
If \code{FALSE}, uses the argument \code{cf}
to generate ISBN.}

\item{doi}{Character vector
A vector of Digital Object Identifiers.}

\item{cf}{R Object.
Result of \code{lib_metadata_crossref}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Extract ISBN from the result of lib_metadata_crossref.
}
\examples{
# Single DOI
doi <- "10.1007/978-0-387-98141-3"
cf <- lib_metadata_crossref(
  doi = doi,
  par = FALSE
)
lib_extract_isbn_crossref(
  fromdoi = TRUE,
  doi = doi,
  par = FALSE
)
lib_extract_isbn_crossref(
  fromdoi = FALSE,
  cf = cf,
  par = FALSE
)
# Vector of DOIs
doi <- c(
  "https://doi.org/10.1007/978-0-387-71762-3",
  "10.1007/978-0-387-98141-3"
)
cf <- lib_metadata_crossref(
  doi = doi,
  par = FALSE
)
lib_extract_isbn_crossref(
  fromdoi = TRUE,
  doi = doi,
  par = FALSE
)
lib_extract_isbn_crossref(
  fromdoi = FALSE,
  cf = cf,
  par = FALSE
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
