% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_isbn2googlebooksid.R
\name{lib_isbn2googlebooksid}
\alias{lib_isbn2googlebooksid}
\title{ISBN to Google Books ID}
\usage{
lib_isbn2googlebooksid(isbn, par = TRUE, ncores = NULL)
}
\arguments{
\item{isbn}{Character vector.
International Standard Book Number.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified,
defaults to \code{detectCores() - 1}.}
}
\description{
ISBN to Google Books ID
}
\examples{
# Single ISBN
isbn <- "978-0-387-98141-3"
lib_isbn2googlebooksid(
  isbn = isbn,
  par = FALSE
)
# Vector of ISBNs
isbn <- c(
  "978-0-387-98141-3",
  "978-0-387-71762-3"
)
lib_isbn2googlebooksid(
  isbn = isbn,
  par = FALSE
)
}
