% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_metadata_calibre.R
\name{lib_metadata_calibre}
\alias{lib_metadata_calibre}
\title{Extract Calibre Metadata}
\usage{
lib_metadata_calibre(isbn, par = TRUE, ncores = NULL)
}
\arguments{
\item{isbn}{Character vector.
International Standard Book Number.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\description{
Extracts metadata using Calibre's \code{fetch-ebook-metadata} command
with the argument \code{--identifier isbn:ITEM_ISBN}.
}
\examples{
isbn <- c(
  "978-0-387-98141-3",
  "978-0-387-71762-3"
)
lib_metadata_calibre(
  isbn = isbn,
  par = FALSE
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
