% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_metadata_crossref.R
\name{lib_metadata_crossref}
\alias{lib_metadata_crossref}
\title{Extract Crossref Metadata}
\usage{
lib_metadata_crossref(doi, par = TRUE, ncores = NULL)
}
\arguments{
\item{doi}{Character vector
A vector of Digital Object Identifiers.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified,
defaults to \code{detectCores() - 1}.}
}
\description{
Extracts Crossref metadata from the Crossref API
using the item's Digital Object Identifier (DOI).
}
\examples{
# Single DOI
doi <- "https://doi.org/10.1016/j.adolescence.2014.04.012"
lib_metadata_crossref(
  doi = doi,
  par = FALSE
)
doi <- "10.1016/j.adolescence.2014.04.012"
lib_metadata_crossref(
  doi = doi,
  par = FALSE
)
# Vector of DOIs
doi <- c(
  "https://doi.org/10.1016/j.adolescence.2014.04.012",
  "https://doi.org/10.1007/s11469-015-9612-8"
)
lib_metadata_crossref(
  doi = doi,
  par = FALSE
)
doi <- c(
  "10.1016/j.adolescence.2014.04.012",
  "10.1007/s11469-015-9612-8"
)
lib_metadata_crossref(
  doi = doi,
  par = FALSE
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
