% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_metadata_lc.R
\name{lib_metadata_lc}
\alias{lib_metadata_lc}
\title{Extract Library of Congress Metadata.}
\usage{
lib_metadata_lc(lccn, par = TRUE, ncores = NULL)
}
\arguments{
\item{lccn}{Character vector.
Library of Congress Control Number (LCCN)}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\description{
Extracts Library of Congress metadata from the LC API.
}
\examples{
lccn <- c(
  "2007925720",
  "2009928510"
)
lib_metadata_lc(
  lccn = lccn,
  par = FALSE
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
