% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_metadata_openlib.R
\name{lib_metadata_openlib}
\alias{lib_metadata_openlib}
\title{Extract Open Library Metadata}
\usage{
lib_metadata_openlib(isbn, par = TRUE, ncores = NULL)
}
\arguments{
\item{isbn}{Character vector.
International Standard Book Number.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Extract Open Library Metadata
}
\examples{
isbn <- c(
  "978-0-387-98141-3",
  "978-0-387-71762-3"
)
lib_metadata_openlib(
  isbn = isbn,
  par = FALSE
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
