% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_remove_doi_http.R
\name{lib_remove_doi_http}
\alias{lib_remove_doi_http}
\title{Remove DOI HTTP.}
\usage{
lib_remove_doi_http(doi)
}
\arguments{
\item{doi}{Character vector.
A vector of Digital Object Identifiers.}
}
\description{
Removes DOI HTTP using the following patters:
\itemize{
\item \verb{http://dx.doi.org/}
\item \verb{http://doi.org/}
\item \verb{https://dx.doi.org/}
\item \verb{https://doi.org/}
}
}
\examples{
# Single DOI
doi <- "https://doi.org/10.1016/j.adolescence.2014.04.012"
lib_remove_doi_http(doi = doi)
# Vector of DOIs
doi <- c(
  "https://doi.org/10.1016/j.adolescence.2014.04.012",
  "https://doi.org/10.1007/s11469-015-9612-8",
  "https://doi.org/10.1016/j.addbeh.2014.10.002",
  "https://doi.org/10.1007/s10566-014-9275-9",
  "https://doi.org/10.1037/a0039101",
  "https://doi.org/10.1016/j.paid.2016.10.059",
  "https://doi.org/10.1016/j.addbeh.2016.03.026",
  "https://doi.org/10.1080/10826084.2019.1658784",
  "https://doi.org/10.1016/j.addbeh.2018.11.045"
)
lib_remove_doi_http(doi = doi)
}
\author{
Ivan Jacob Agaloos Pesigan
}
