% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_remove_doi_prefix.R
\name{lib_remove_doi_prefix}
\alias{lib_remove_doi_prefix}
\title{Remove DOI Prefix.}
\usage{
lib_remove_doi_prefix(doi)
}
\arguments{
\item{doi}{Character vector.
A vector of Digital Object Identifiers.}
}
\description{
Removes DOI HTTP
(see \code{\link[=lib_remove_doi_http]{lib_remove_doi_http()}})
and
prefix,
that is, the first 6 digits of a DOI.
For example, in the case of
\code{"https://doi.org/10.1016/j.adolescence.2014.04.012"}
both
\code{"https://doi.org/"}
and \code{"10.1016/"}
are removed resulting in
\code{"j.adolescence.2014.04.012"}.
}
\details{
For more information on DOIs and matching regular expressions see
\href{https://www.crossref.org/blog/dois-and-matching-regular-expressions/}{DOI regex}.

Regular expressions
\itemize{
\item \verb{/^10.\\d\{4,9\}/[-._;()/:A-Z0-9]+$/i}
\item \verb{/^10.1002/[^\\s]+$/i}
\item \verb{/^10.\\d\{4\}/\\d+-\\d+X?(\\d+)\\d+<[\\d\\w]+:[\\d\\w]*>\\d+.\\d+.\\w+;\\d$/i}
\item \verb{/^10.1021/\\w\\w\\d++$/i}
\item \verb{/^10.1207/[\\w\\d]+\\&\\d+_\\d+$/i}
}
}
\examples{
# Single DOI
doi <- "https://doi.org/10.1016/j.adolescence.2014.04.012"
lib_remove_doi_prefix(doi = doi)
doi <- "10.1016/j.adolescence.2014.04.012"
lib_remove_doi_prefix(doi = doi)
# Vector of DOIs
doi <- c(
  "https://doi.org/10.1016/j.adolescence.2014.04.012",
  "https://doi.org/10.1007/s11469-015-9612-8",
  "https://doi.org/10.1016/j.addbeh.2014.10.002",
  "https://doi.org/10.1007/s10566-014-9275-9",
  "https://doi.org/10.1037/a0039101",
  "https://doi.org/10.1016/j.paid.2016.10.059",
  "https://doi.org/10.1016/j.addbeh.2016.03.026",
  "https://doi.org/10.1080/10826084.2019.1658784",
  "https://doi.org/10.1016/j.addbeh.2018.11.045"
)
lib_remove_doi_prefix(doi = doi)
doi <- c(
  "10.1016/j.adolescence.2014.04.012",
  "10.1007/s11469-015-9612-8",
  "10.1016/j.addbeh.2014.10.002",
  "10.1007/s10566-014-9275-9",
  "10.1037/a0039101",
  "10.1016/j.paid.2016.10.059",
  "10.1016/j.addbeh.2016.03.026",
  "10.1080/10826084.2019.1658784",
  "10.1016/j.addbeh.2018.11.045"
)
lib_remove_doi_prefix(doi = doi)
}
\author{
Ivan Jacob Agaloos Pesigan
}
