% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_remove_isbn_dashes.R
\name{lib_remove_isbn_dashes}
\alias{lib_remove_isbn_dashes}
\title{Remove ISBN Dashes.}
\usage{
lib_remove_isbn_dashes(isbn)
}
\arguments{
\item{isbn}{Character vector.
International Standard Book Number.}
}
\description{
Remove ISBN Dashes.
}
\examples{
# Single ISBN
isbn <- "978-3-030-29184-6"
lib_remove_isbn_dashes(isbn = isbn)
# Vector of ISBNs
isbn <- c(
  "978-0-387-98141-3",
  "978-0-387-71762-3"
)
lib_remove_isbn_dashes(isbn = isbn)
}
\author{
Ivan Jacob Agaloos Pesigan
}
