% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_springer_calibredb_add.R
\name{lib_springer_calibredb_add}
\alias{lib_springer_calibredb_add}
\title{Add Springer Books to Calibre}
\usage{
lib_springer_calibredb_add(doi, dir, calibre_dir)
}
\arguments{
\item{doi}{Character vector.
A vector of Digital Object Identifiers.}

\item{dir}{Character string.
Directory where Springer Books are located.}

\item{calibre_dir}{Character string.
Directory where Calibre's Library Database is located.}
}
\description{
Add Springer Books to Calibre
}
\author{
Ivan Jacob Agaloos Pesigan
}
