% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_springer_download.R
\name{lib_springer_download}
\alias{lib_springer_download}
\title{Download Resources from link.springer.com.}
\usage{
lib_springer_download(
  dir,
  doi = NULL,
  chkfiles = TRUE,
  args = "-nc",
  remove_files = TRUE,
  type = c("pdf", "epub", "both"),
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{dir}{Character string.
Directory which contains target files.}

\item{doi}{Character vector.
A vector of Digital Object Identifiers.}

\item{chkfiles}{Logical.
Check available files in \code{dir}.
If \code{TRUE},
the download queue will EXCLUDE dois
from available files.}

\item{args}{Character string.
Arguments to pass to \code{wget}.}

\item{remove_files}{Logical.
Remove files that do not match the specified file type.
BE CAREFUL WITH THIS OPTION
AS FILES CAN BE DELETED FROM YOUR SYSTEM.}

\item{type}{Character string.
Type of ebooks to download.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified,
defaults to \code{detectCores() - 1}.}
}
\description{
Downloads \code{pdf} and \code{epub} files
from \code{link.springer.com}
using \code{wget}.
It is assumed that you have access
to the resources via your library.
It also assumes that \code{wget}
is installed in your system.
}
\examples{
\dontrun{
# Single DOI
doi <- "10.1007/978-0-387-98141-3"
lib_springer_download(
  dir = getwd(),
  doi = doi,
  chkfiles = TRUE,
  type = "both",
  par = FALSE
)
# Vector of DOIs
doi <- c(
  "https://doi.org/10.1007/978-0-387-71762-3",
  "10.1007/978-0-387-98141-3"
)
lib_springer_download(
  dir = getwd(),
  doi = doi,
  chkfiles = TRUE,
  type = "both",
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
