% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_springer_files.R
\name{lib_springer_files}
\alias{lib_springer_files}
\title{Create a list of available Springer books.}
\usage{
lib_springer_files(dir)
}
\arguments{
\item{dir}{Character string.
Directory which contains target files.}
}
\description{
Create a list of available Springer books.
}
\author{
Ivan Jacob Agaloos Pesigan
}
