% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_springer_link.R
\name{lib_springer_link}
\alias{lib_springer_link}
\title{Generate Spinger Links from DOI.}
\usage{
lib_springer_link(doi = "10.1007/978-0-387-79054-1", PDF = TRUE)
}
\arguments{
\item{doi}{Character vector.
A vector of Digital Object Identifiers.}

\item{PDF}{Logical.
If \code{TRUE}, returns links for \code{PDF} files.
If \code{FALSE}, returns links for \code{EPUB} files.}
}
\description{
Generate Spinger Links from DOI.
}
\examples{
# Single DOI
doi <- "10.1007/978-0-387-98141-3"
lib_springer_link(
  doi = doi,
  PDF = TRUE
)
lib_springer_link(
  doi = doi,
  PDF = FALSE
)
# Vector of DOIs
doi <- c(
  "https://doi.org/10.1007/978-0-387-71762-3",
  "10.1007/978-0-387-98141-3"
)
lib_springer_link(
  doi = doi,
  PDF = TRUE
)
lib_springer_link(
  doi = doi,
  PDF = FALSE
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
