% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_wiley_doi2isbn.R
\name{lib_wiley_doi2isbn}
\alias{lib_wiley_doi2isbn}
\title{DOI to ISBN for Wiley Books}
\usage{
lib_wiley_doi2isbn(doi, par = TRUE, ncores = NULL)
}
\arguments{
\item{doi}{Character vector.
A vector of Digital Object Identifiers.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified,
defaults to \code{detectCores() - 1}.}
}
\description{
DOI to ISBN for Wiley Books
}
\examples{
\dontrun{
doi <- c(
  "10.1002/9780470515075",
  "10.1002/9780470721896"
)
lib_wiley_doi2isbn(
  doi = doi,
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
