% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_springer_books_doi.R
\docType{data}
\name{springer_books_doi}
\alias{springer_books_doi}
\title{Springer Books DOI}
\format{
A vector of Springer books DOIs.
}
\usage{
springer_books_doi
}
\description{
Springer Books DOI
}
\keyword{datasets}
